extension: Move ecssd into extension/observer/ecsobeserver
There is no plan to reuse it in other aws related components
