+++ extension/observer/ecsobserver/service.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package ecssd
+package ecsobserver
 
 type ServiceConfig struct {
 	CommonExporterConfig `mapstructure:",squash" yaml:",inline"`
